[cleanup] simplify pass: drop dead fallback, narration comments, no-op cast
Review surfaced three unambiguous wins:

- ``STULayer.forward`` had a fallback that built ``attn_func`` locally
  when the caller passed ``None`` and SLA was enabled on the layer.
  STUStack.forward already hoists ``attn_func`` construction to stack
  scope and always supplies it when needed, so the fallback is never
  exercised. Removed the dead branch and updated the kwarg docstring
  to say STUStack is the sole constructor.

- ``apply_stu_truncation`` carried "Step 1: / Step 2: / Step 3:"
  narration comments above the three primitive calls. The function's
  docstring already explains the three-part slice strategy; the
  in-body narration just restates what the next line does. Dropped.

- ``_decode_attn_func_to_mask`` called ``.to(torch.int32)`` on the
  attn_func tensor after reshape; attn_func is documented as int32 in
  the function's Args section, so the cast is a no-op allocation on
  the hot mask-decode path. Dropped.

+++ tzrec/ops/_pytorch/pt_hstu_attention.py
@@ -147,7 +147,7 @@ def _decode_attn_func_to_mask(
     torch._assert(three == 3, "attn_func must have shape (H, 3, total_q)")
     # Fold (H, 3) into channels so we can use jagged_to_padded_dense with
     # the (total_q, C) 2D layout; unfold after padding.
-    padded_flat = attn_func.permute(2, 0, 1).reshape(total_q, H * 3).to(torch.int32)
+    padded_flat = attn_func.permute(2, 0, 1).reshape(total_q, H * 3)
     padded = torch.ops.fbgemm.jagged_to_padded_dense(
         values=padded_flat,
         offsets=[seq_offsets],
